[fix] add int for ln 35 adn ln 36

+++ math/0x04-convolutions_and_pooling/1-convolve_grayscale_same.py
@@ -31,9 +31,9 @@ def convolve_grayscale_same(images, kernel):
     pad_w = int((kw - 1) / 2)
 
     if kh % 2 == 0:
-        pad_h = kh / 2
+        pad_h = int(kh / 2)
     if kw % 2 == 0:
-        pad_w = kw / 2
+        pad_w = int(kw / 2)
 
     image_pad = np.pad(images, pad_width=((0, 0),
                                           (pad_h, pad_h), (pad_w, pad_w)),
